Allow receiving of float for event values, update python cli gen

+++ event/persister/event.go
@@ -19,8 +19,8 @@ func SaveEvent(ctx context.Context, event *api.NewMeasurementEvent) error {
 	_, err := db.ExecContext(
 		ctx,
 		`INSERT INTO PeripheralEvents
-		(Id, PeripheralId, DeploymentId, Value, Timestamp)
-		VALUES (UUID(), ?, ?, ?, ?);`,
+        (Id, PeripheralId, DeploymentId, Value, Timestamp)
+        VALUES (UUID(), ?, ?, ?, ?);`,
 		&event.PeripheralId,
 		&event.DeploymentId,
 		&event.Value,
@@ -47,9 +47,9 @@ func FilterEvents(ctx context.Context, request *api.MeasurementEventFilterReques
 	events, err := ScanEvents(db.QueryContext(
 		ctx,
 		`SELECT * FROM PeripheralEvents WHERE
-		Timestamp BETWEEN ? AND ?
-		AND DeploymentId = ?
-		AND PeripheralId = ?;`,
+        Timestamp BETWEEN ? AND ?
+        AND DeploymentId = ?
+        AND PeripheralId = ?;`,
 		starttime,
 		endtime,
 		&request.DeploymentId,
@@ -67,12 +67,12 @@ func FilterEvents(ctx context.Context, request *api.MeasurementEventFilterReques
 
 func initTable(ctx context.Context, pool *sql.DB) error {
 	_, peripheralEventsTableCreateErr := pool.ExecContext(ctx, `CREATE TABLE IF NOT EXISTS PeripheralEvents (
-Id varchar(36) PRIMARY KEY NOT NULL,
-		PeripheralId varchar(36) NOT NULL,
-		DeploymentId varchar(36) NOT NULL,
-		Value smallint NOT NULL,
-		Timestamp timestamp NOT NULL
-	    );`,
+        Id varchar(36) PRIMARY KEY NOT NULL,
+        PeripheralId varchar(36) NOT NULL,
+        DeploymentId varchar(36) NOT NULL,
+        Value float NOT NULL,
+        Timestamp timestamp NOT NULL
+        );`,
 	)
 	if peripheralEventsTableCreateErr != nil {
 		log.Printf(
